# Perf improvements for large channels