--- a/server/src/dao/entities/BaseEntity.ts
+++ b/server/src/dao/entities/BaseEntity.ts
@@ -1,5 +1,6 @@
 import {
   Entity,
+  OptionalProps,
   BaseEntity as OrmBaseEntity,
   PrimaryKey,
   Property,
@@ -8,12 +9,14 @@ import { v4 } from 'uuid';
 
 @Entity({ abstract: true })
 export abstract class BaseEntity extends OrmBaseEntity {
+  [OptionalProps]?: 'createdAt' | 'updatedAt';
+
   @PrimaryKey()
   uuid: string = v4();
 
-  @Property({ type: 'datetime', onCreate: () => new Date() })
-  createdAt?: Date = new Date();
+  @Property({ type: 'datetime' })
+  createdAt: Date = new Date();
 
   @Property({ type: 'datetime', onUpdate: () => new Date() })
-  updatedAt?: Date = new Date();
+  updatedAt: Date = new Date();
 }