On my 2018 Macbook Pro, I'm seeing updates for channels with >2k programs take 2-3s and retrieving the lineup take 1.75-2s. There's definitely more room for improvement.

Includes:
* Batching inserts to external IDs
* Only saving critical external IDs (plex rating key) synchronously on
  lineup update. Non-critical external ID inserts are pushed to
  background tasks
* Program grouping / hierarchy updating has been pushed to a background
  task
* Optimized verison of content program converter function that is
  synchronous. This is much more performant for 1000s of items,
  specifically when no async actions (selecting extra data) is
  necessary. We saw a ~89% reduction in time spent here (>1s to ~30ms)
* A few places were selecting all channel programs when they only needed
  to select the channel
* Only select program IDs at the outset of lineup update, instead of all
  program details. We only need the IDs for relation diffing

Still some things we haven't solved:

1. There is non-trivial overhead in the mikro-orm ORM framework to
   mapping 1000s of selected entities. In a 2k program channel case,
   when loading all necessary entities (relations), we can see ~6k
   entities loaded by the framework. The select from the DB only takes
   about 800ms, but the entity mapping step can take >3s in this case.
   One solution here is to use a simpler library for these super large
   selects (kysely?)
2. It's probably overkill to have both Zodios on the client and then Zod
   checking incoming types on the server. On huge request, this can add
   ~100ms or more (server side) to requests as Zod validates incoming
   requests against the schema
3. We should think about replacing Zod on the server side with JSON
   Schema. There are converters out there. We have a lot invested in
   Zod, so a converter would be a good first step.
4. There's clearly some I/O contention happening in certain situations
   ... background tasks that query DB or Plex, getting responses to flow, logging, etc. I think most of it is DB-related.
5. Unclear if there is any actually _fast_ way to insert the amount of
   data we are currently generating for large channels.
6. I purposely broke custom show resolution on channel lineup retrieval because it was super slow and I didnt have a better solution. Need to fix it